requirements.txt
# Core Framework
fastapi==0.109.0
uvicorn[standard]==0.27.0
python-multipart==0.0.6

# Database
sqlalchemy==2.0.25
psycopg2-binary==2.9.9
alembic==1.13.1

# Vector Database
qdrant-client==1.7.3
sentence-transformers==2.3.1

# Authentication & Security
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-dotenv==1.0.0
bcrypt==4.1.2
email-validator==2.1.0

# OAuth & Social Media APIs
httpx==0.26.0
tweepy==4.14.0
facebook-sdk==3.1.0
linkedin-api==2.2.0
google-auth==2.27.0
google-auth-oauthlib==1.2.0
google-auth-httplib2==0.2.0
google-api-python-client==2.116.0

# File Processing
PyPDF2==3.0.1
python-docx==1.1.0
python-magic==0.4.27

# Utilities
pydantic==2.5.3
pydantic-settings==2.1.0
python-dateutil==2.8.2
requests==2.31.0

# AI & Embeddings
openai==1.12.0
cohere==4.47

# Background Tasks
celery==5.3.6
redis==5.0.1
